+++ Fluent.Ribbon/Controls/DropDownButton.cs
@@ -3,8 +3,10 @@ namespace Fluent
 {
     using System;
     using System.Collections;
+    using System.Collections.Generic;
     using System.Diagnostics;
     using System.Threading;
+    using System.Threading.Tasks;
     using System.Windows;
     using System.Windows.Controls;
     using System.Windows.Controls.Primitives;
@@ -38,6 +40,8 @@ namespace Fluent
         private ScrollViewer scrollViewer;
 
         private UIElement buttonBorder;
+
+        private readonly Stack<WeakReference> openMenuItems = new Stack<WeakReference>();
 
         #endregion
 
@@ -369,6 +373,9 @@ namespace Fluent
 
             this.Loaded += this.OnLoaded;
             this.Unloaded += this.OnUnloaded;
+
+            this.AddHandler(System.Windows.Controls.MenuItem.SubmenuOpenedEvent, new RoutedEventHandler(this.OnSubmenuOpened));
+            this.AddHandler(System.Windows.Controls.MenuItem.SubmenuClosedEvent, new RoutedEventHandler(this.OnSubmenuClosed));
         }
 
         private void OnLoaded(object sender, RoutedEventArgs e)
@@ -512,15 +519,19 @@ namespace Fluent
                 && this.resizeBothThumb.IsMouseOver == false
                 && this.resizeVerticalThumb.IsMouseOver == false)
             {
-                e.Handled = false;
-
                 // Note: get outside thread to prevent exceptions (it's a dependency property after all)
-                var closePopupOnMouseDownDelay = this.ClosePopupOnMouseDownDelay;
+                var timespan = this.ClosePopupOnMouseDownDelay;
 
                 // Ugly workaround, but use a timer to allow routed event to continue
-                System.Threading.Tasks.Task.Factory.StartNew(() =>
+#if NET40
+                Task.Factory.StartNew(() =>
                 {
-                    Thread.Sleep(closePopupOnMouseDownDelay);
+                    Thread.Sleep(timespan);
+#else
+                Task.Factory.StartNew(async () =>
+                {
+                    await Task.Delay(timespan);
+#endif
 
                     this.RunInDispatcherAsync(() => this.IsDropDownOpen = false);
                 });
@@ -558,7 +569,7 @@ namespace Fluent
 
                         var container = this.ItemContainerGenerator.ContainerFromIndex(0);
 
-                        NavigateToContainer(container);
+                        NavigateToContainer(container, FocusNavigationDirection.Down);
 
                         handled = true;
                     }
@@ -573,7 +584,7 @@ namespace Fluent
 
                         var container = this.ItemContainerGenerator.ContainerFromIndex(this.Items.Count - 1);
 
-                        NavigateToContainer(container);
+                        NavigateToContainer(container, FocusNavigationDirection.Up);
 
                         handled = true;
                     }
@@ -604,7 +615,7 @@ namespace Fluent
             base.OnKeyDown(e);
         }
 
-        private static void NavigateToContainer(DependencyObject container)
+        private static void NavigateToContainer(DependencyObject container, FocusNavigationDirection focusNavigationDirection = FocusNavigationDirection.Down)
         {
             var element = container as FrameworkElement;
 
@@ -619,12 +630,7 @@ namespace Fluent
             }
             else
             {
-                var predicted = element.PredictFocus(FocusNavigationDirection.Down);
-
-                if (predicted is MenuBase == false)
-                {
-                    element.MoveFocus(new TraversalRequest(FocusNavigationDirection.Down));
-                }
+                element.MoveFocus(new TraversalRequest(focusNavigationDirection));
             }
         }
 
@@ -643,12 +649,6 @@ namespace Fluent
         public virtual KeyTipPressedResult OnKeyTipPressed()
         {
             this.IsDropDownOpen = true;
-
-            if (this.DropDownPopup?.Child != null)
-            {
-                Keyboard.Focus(this.DropDownPopup.Child);
-                this.DropDownPopup.Child.MoveFocus(new TraversalRequest(FocusNavigationDirection.First));
-            }
 
             return new KeyTipPressedResult(true, true);
         }
@@ -748,16 +748,30 @@ namespace Fluent
             }
         }
 
-        // Handles drop down closed
-        private void OnDropDownClosed()
-        {
-            this.DropDownClosed?.Invoke(this, EventArgs.Empty);
-        }
-
-        // Handles drop down opened
-        private void OnDropDownOpened()
+        /// <summary>
+        /// Called when drop down opened.
+        /// </summary>
+        protected virtual void OnDropDownOpened()
         {
             this.DropDownOpened?.Invoke(this, EventArgs.Empty);
+        }
+
+        /// <summary>
+        /// Called when drop down closed.
+        /// </summary>
+        protected virtual void OnDropDownClosed()
+        {
+            foreach (var openMenuItem in this.openMenuItems.ToArray())
+            {
+                if (openMenuItem.IsAlive)
+                {
+                    ((System.Windows.Controls.MenuItem)openMenuItem.Target).IsSubmenuOpen = false;
+                }
+            }
+
+            this.openMenuItems.Clear();
+
+            this.DropDownClosed?.Invoke(this, EventArgs.Empty);
         }
 
         #endregion
@@ -865,12 +879,7 @@ namespace Fluent
 
         #endregion
 
-        /// <summary>
-        /// Gets an enumerator for the logical child objects of the <see cref="T:System.Windows.Controls.ItemsControl"/> object.
-        /// </summary>
-        /// <returns>
-        /// An enumerator for the logical child objects of the <see cref="T:System.Windows.Controls.ItemsControl"/> object. The default is null.
-        /// </returns>
+        /// <inheritdoc />
         protected override IEnumerator LogicalChildren
         {
             get
@@ -891,5 +900,26 @@ namespace Fluent
                 }
             }
         }
+
+        #region MenuItem workarounds
+
+        private void OnSubmenuOpened(object sender, RoutedEventArgs e)
+        {
+            var menuItem = e.OriginalSource as MenuItem;
+            if (menuItem != null)
+            {
+                this.openMenuItems.Push(new WeakReference(menuItem));
+            }
+        }
+
+        private void OnSubmenuClosed(object sender, RoutedEventArgs e)
+        {
+            if (this.openMenuItems.Count > 0)
+            {
+                this.openMenuItems.Pop();
+            }
+        }
+
+        #endregion MenuItem workarounds
     }
 }