+++ Fluent.Ribbon/Localization/Languages/Korean.cs
@@ -5,7 +5,7 @@ namespace Fluent.Localization.Languages
     [RibbonLocalization("Korean", "ko")]
     public class Korean : RibbonLocalizationBase
     {
-        public override string Automatic { get; } = "자동 번역";
+        public override string Automatic { get; } = "자동";
         public override string BackstageButtonKeyTip { get; } = FallbackLocalization.BackstageButtonKeyTip;
         public override string BackstageButtonText { get; } = "파일";
         public override string CustomizeStatusBar { get; } = "상태 표시줄 사용자 지정";